pull out some reusable testing tools for shallow rendering
