+++ test/ReactViews/FeatureInfoPanelSpec.jsx
@@ -3,12 +3,15 @@
 /*global require,expect*/
 // import knockout from 'terriajs-cesium/Source/ThirdParty/knockout';
 import React from 'react';
-import ReactTestUtils from 'react-addons-test-utils';
-import {findWithType} from 'react-shallow-testutils';
+import {findWithType, findAllWithType} from 'react-shallow-testutils';
+import {getShallowRenderedOutput} from './MoreShallowTools';
 
 // import Entity from 'terriajs-cesium/Source/DataSources/Entity';
 
+import Entity from 'terriajs-cesium/Source/DataSources/Entity';
+
 import FeatureInfoPanel  from '../../lib/ReactViews/FeatureInfo/FeatureInfoPanel';
+import FeatureInfoSection from '../../lib/ReactViews/FeatureInfo/FeatureInfoSection';
 import Loader from '../../lib/ReactViews/Loader';
 import PickedFeatures from '../../lib/Map/PickedFeatures';
 import runLater from '../../lib/Core/runLater';
@@ -20,13 +23,7 @@ import ViewState from '../../lib/ReactViewModels/ViewState';
 //     separator = (Intl.NumberFormat().format(1000)[1]);
 // }
 
-function getShallowRenderedOutput(jsx) {
-    const renderer = ReactTestUtils.createRenderer();
-    renderer.render(jsx);
-    return renderer.getRenderOutput();
-}
-
-describe('FeatureInfoPanel-jsx', function() {
+describe('FeatureInfoPanel', function() {
 
     let terria;
     // let feature;
@@ -39,14 +36,6 @@ describe('FeatureInfoPanel-jsx', function() {
         viewState = new ViewState({
             terria: terria
         });
-        // const properties = {
-        //     'Foo': 'bar',
-        //     'moo': 'd"e"r,p'
-        // };
-        // feature = new Entity({
-        //     name: 'Bar',
-        //     properties: properties
-        // });
     });
 
     it('does not have isVisible class when viewState.featureInfoPanelIsVisible not set', function() {
@@ -71,26 +60,27 @@ describe('FeatureInfoPanel-jsx', function() {
         expect(findWithType(result, Loader)).toBeDefined();
     });
 
+    // This test won't work for two reasons:
+    //   - the behaviour it tests occurs in ComponentDidMount
+    //   - FeatureInfoPanel doesn't have FeatureInfoSections - there is a FeatureInfoCatalogItem layer in between.
+    //
+    // it('shows an open section even if none have any info', function() {
+    //     const feature1 = new Entity({
+    //         name: 'Foo'
+    //     });
+    //     const feature2 = new Entity({
+    //         name: 'Bar'
+    //     });
+    //     var pickedFeatures = new PickedFeatures();
+    //     pickedFeatures.allFeaturesAvailablePromise = runLater(function() {
+    //         pickedFeatures.features = [feature1, feature2];
+    //     });
+    //     terria.pickedFeatures = pickedFeatures;
+    //     const panel = <FeatureInfoPanel terria={terria} viewState={viewState}/>;
+    //     const result = getShallowRenderedOutput(panel);
+    //     const sections = findAllWithType(result, FeatureInfoSection);
+    //     expect(sections.length).toEqual(2);
+    //     expect(sections[0].props.isOpen).toBe(true);
+    // });
+
 });
-
-// function treeContainsText(tree, s) {
-//     if (tree.props.children === undefined) {
-//         return false;
-//     }
-//     if (tree.props.children.props !== undefined) {
-//         // A single child is sometimes present in children without being in an array.
-//         return treeContainsText(tree.props.children, s)
-//     }
-//     if (!Array.isArray(tree.props.children)) {
-//         // Text is presented as a child, with no object around it.
-//         console.log(tree.props.children);
-//         return tree.props.children.indexOf(s) >= 0;
-//     }
-//     for (var i = 0; i < tree.props.children.length; ++i) {
-//         if (treeContainsText(tree.props.children[i], s)) {
-//             return true;
-//         }
-//     }
-
-//     return false;
-// }
